feat: get question bank from api request
