+++ .env.sample
@@ -1,0 +1,1 @@
+VITE_API_PREFIX=example
