+++ .gitignore
@@ -14,6 +14,7 @@ dist-ssr
 
 coverage
 .eslintcache
+.env
 
 # Editor directories and files
 .vscode/*
