--- src/backups/Leaderboard.ts
@@ -1,9 +1,0 @@
-import { LeaderboardRow } from "../utils/Utils.types";
-
-export const initLeaderboardRows: LeaderboardRow[] = [
-    { rank: 1, player: "Infinity", score: 450, level: "Sage" },
-    { rank: 2, player: "Zero", score: 420, level: "Sage" },
-    { rank: 3, player: "IKnowALot", score: 380, level: "Expert" },
-    { rank: 4, player: "RandomPlayer", score: 350, level: "Expert" },
-    { rank: 5, player: "Fleia", score: 320, level: "Expert" },
-];
